+++ src/libaatm/src/ATMWVRMeasurement.cpp
@@ -43,7 +43,7 @@ WVRMeasurement::WVRMeasurement(const Angle &elevation,
   v_measuredSkyBrightness_ = measuredSkyBrightness;
   retrievedWaterVaporColumn_ = Length(-999, "mm"); // !< Retrieved zenith water vapor column for each event
   sigma_fittedSkyBrightness_ = Temperature(-999, "K"); // !< Sigma on the fitted sky brightness temperatures (average sigma over the WVR channels for each event).
-  for(unsigned int i = 0; i < v_measuredSkyBrightness_.size(); i++) {
+  for(size_t i = 0; i < v_measuredSkyBrightness_.size(); i++) {
     v_fittedSkyBrightness_.push_back(Temperature(-999, "K"));
   } // !< Fitted sky brightness temperatures over all WVR channels for each event
 }
